[REFACT] : onDragEnd 프롭 삭제, SortableHeader 프롭생성
